Add functional configuration for Connect to permit custom clients
This is required if we want to connect to the HTTPS listener running on
Neo4j when running with a self-signed certificate. In this circumstance
we need to configure a client using the corresponding self-signed cert,
so we need to be able to pass in this client for the neoism driver to
use.

We added a little unit test to verify that the custom client is set on
the driver.