Attempt to resolve local race condition in tests.
Locally I was seeing tests fail intermittently due to the index created
in `TestCreateIndexes` not being deleted by the deferred clean up
indexes call, meaning that a later test that asserted on the number of
indexes would fail.

This change drops the index inline using `idx.Drop()`, which seemed to
pass tests reliably for me.